+++ packages/charts/src/index.ts
@@ -43,14 +43,14 @@ export {
 export { GeometryValue, BandedAccessorType } from './utils/geometry';
 export { LegendPath, LegendPathElement } from './state/actions/legend';
 export { CategoryKey } from './common/category';
-export {
-  Config as PartitionConfig,
-  FillLabelConfig as PartitionFillLabel,
-  PartitionLayout,
-} from './chart_types/partition_chart/layout/types/config_types';
+export { FillLabelConfig as PartitionFillLabel, PartitionStyle } from './utils/themes/partition';
+export { PartitionLayout } from './chart_types/partition_chart/layout/types/config_types';
 export { Layer as PartitionLayer } from './chart_types/partition_chart/specs/index';
 export * from './chart_types/goal_chart/specs/index';
 export * from './chart_types/wordcloud/specs/index';
+
+// TODO: Remove deprecated config types. Need for eui alignment
+export * from './chart_types/partition_chart/layout/types/config';
 
 export {
   Accessor,
@@ -72,6 +72,9 @@ export * from './utils/themes/theme_common';
 export { LIGHT_THEME } from './utils/themes/light_theme';
 export { DARK_THEME } from './utils/themes/dark_theme';
 
+// wordcloud
+export { WordcloudViewModel } from './chart_types/wordcloud/layout/types/viewmodel_types';
+
 // partition
 export * from './chart_types/partition_chart/layout/types/viewmodel_types';
 export * from './chart_types/partition_chart/layout/utils/group_by_rollup';
@@ -79,9 +82,6 @@ export { AnimKeyframe } from './chart_types/partition_chart/layout/types/config_
 
 // heatmap
 export { Cell } from './chart_types/heatmap/layout/types/viewmodel_types';
-export { SizeRatio, TimeMs } from './common/geometry';
-export { TextAlign, TextBaseline } from './common/text_utils';
-export { Config as HeatmapConfig } from './chart_types/heatmap/layout/types/config_types';
 export { ColorBand, HeatmapBandsColorScale } from './chart_types/heatmap/specs/heatmap';
 
 // utilities
@@ -106,7 +106,7 @@ export {
 export { DataGenerator } from './utils/data_generators/data_generator';
 export * from './utils/themes/merge_utils';
 export * from './utils/use_legend_action';
-export { MODEL_KEY } from './chart_types/partition_chart/layout/config';
+export { MODEL_KEY, defaultPartitionValueFormatter } from './chart_types/partition_chart/layout/config';
 export { LegendStrategy } from './chart_types/partition_chart/layout/utils/highlighted_geoms';
 export { Pixels, Ratio } from './common/geometry';
 export { AdditiveNumber } from './utils/accessor';