+++ storybook/stories/heatmap/heatmap.stories.tsx
@@ -14,3 +14,4 @@ export { Example as basic } from './1_basic.story';
 export { Example as time } from './3_time.story';
 export { Example as categorical } from './2_categorical.story';
 export { Example as timeSnap } from './4_test_time_snap.story';
+export { Example as theming } from './5_theming.story';