+++ storybook/stories/sunburst/32_custom_tooltip.story.tsx
@@ -9,8 +9,16 @@
 import { boolean } from '@storybook/addon-knobs';
 import React from 'react';
 
-import { Chart, Datum, Partition, PartitionLayout, Settings, CustomTooltip as CT } from '@elastic/charts';
-import { config } from '@elastic/charts/src/chart_types/partition_chart/layout/config';
+import {
+  Chart,
+  Datum,
+  Partition,
+  PartitionLayout,
+  Settings,
+  CustomTooltip as CT,
+  PartialTheme,
+  defaultPartitionValueFormatter,
+} from '@elastic/charts';
 import { mocks } from '@elastic/charts/src/mocks/hierarchical';
 
 import { useBaseTheme } from '../../use_base_theme';
@@ -32,6 +40,25 @@ const CustomTooltip: CT = ({ values }) => (
   </div>
 );
 
+const theme: PartialTheme = {
+  chartMargins: { top: 0, left: 0, bottom: 0, right: 0 },
+  partition: {
+    linkLabel: {
+      maxCount: 0,
+      fontSize: 14,
+    },
+    fontFamily: 'Arial',
+    fillLabel: {
+      fontStyle: 'italic',
+    },
+    minFontSize: 1,
+    idealFontSizeJump: 1.1,
+    outerSizeRatio: 0.95,
+    emptySizeRatio: 0,
+    circlePadding: 4,
+  },
+};
+
 export const Example = () => {
   const tooltipOptions = {
     placement: getPlacementKnob('Tooltip placement'),
@@ -41,12 +68,13 @@ export const Example = () => {
   };
   return (
     <Chart>
-      <Settings showLegend legendMaxDepth={1} tooltip={tooltipOptions} baseTheme={useBaseTheme()} />
+      <Settings showLegend legendMaxDepth={1} tooltip={tooltipOptions} theme={theme} baseTheme={useBaseTheme()} />
       <Partition
         id="spec_1"
         data={mocks.sunburst}
+        layout={PartitionLayout.sunburst}
         valueAccessor={(d: Datum) => d.exportVal as number}
-        valueFormatter={(d: number) => `$${config.fillLabel.valueFormatter(Math.round(d / 1000000000))}\u00A0Bn`}
+        valueFormatter={(d: number) => `$${defaultPartitionValueFormatter(Math.round(d / 1000000000))}\u00A0Bn`}
         layers={[
           {
             groupByRollup: (d: Datum) => countryLookup[d.dest].continentCountry.slice(0, 2),
@@ -54,7 +82,7 @@ export const Example = () => {
             fillLabel: {
               fontFamily: 'Impact',
               valueFormatter: (d: number) =>
-                `$${config.fillLabel.valueFormatter(Math.round(d / 1000000000000))}\u00A0Tn`,
+                `$${defaultPartitionValueFormatter(Math.round(d / 1000000000000))}\u00A0Tn`,
             },
             shape: {
               fillColor: (d) =>
@@ -72,25 +100,6 @@ export const Example = () => {
             },
           },
         ]}
-        config={{
-          partitionLayout: PartitionLayout.sunburst,
-          linkLabel: {
-            maxCount: 0,
-            fontSize: 14,
-          },
-          fontFamily: 'Arial',
-          fillLabel: {
-            valueFormatter: (d: number) => `$${config.fillLabel.valueFormatter(Math.round(d / 1000000000))}\u00A0Bn`,
-            fontStyle: 'italic',
-          },
-          margin: { top: 0, bottom: 0, left: 0, right: 0 },
-          minFontSize: 1,
-          idealFontSizeJump: 1.1,
-          outerSizeRatio: 0.95,
-          emptySizeRatio: 0,
-          circlePadding: 4,
-          backgroundColor: 'rgba(229,229,229,1)',
-        }}
       />
     </Chart>
   );