+++ storybook/stories/sunburst/6_pie_chart_linked_labels.story.tsx
@@ -8,33 +8,37 @@
 
 import React from 'react';
 
-import { Chart, Datum, Partition, PartitionLayout, Settings } from '@elastic/charts';
-import { config } from '@elastic/charts/src/chart_types/partition_chart/layout/config';
+import { Chart, Datum, Partition, PartitionLayout, Settings, defaultPartitionValueFormatter } from '@elastic/charts';
 
 import { useBaseTheme } from '../../use_base_theme';
 import { indexInterpolatedFillColor, interpolatorCET2s } from '../utils/utils';
 
 export const Example = () => (
   <Chart>
-    <Settings baseTheme={useBaseTheme()} />
+    <Settings
+      theme={{
+        chartMargins: { top: 0, left: 0, bottom: 0, right: 0 },
+        partition: { linkLabel: { maximumSection: 10000 } },
+      }}
+      baseTheme={useBaseTheme()}
+    />
     <Partition
       id="spec_1"
       data={[
         { sitc1: 'Machinery and transport equipment', exportVal: 5 },
         { sitc1: 'Mineral fuels, lubricants and related materials', exportVal: 4 },
       ]}
+      layout={PartitionLayout.sunburst}
       valueAccessor={(d: Datum) => d.exportVal as number}
-      valueFormatter={(d: number) => `$${config.fillLabel.valueFormatter(Math.round(d))}`}
+      valueFormatter={(d: number) => `$${defaultPartitionValueFormatter(Math.round(d))}`}
       layers={[
         {
           groupByRollup: (d: Datum) => d.sitc1,
-          // nodeLabel: (d: Datum) => d,
           shape: {
             fillColor: indexInterpolatedFillColor(interpolatorCET2s),
           },
         },
       ]}
-      config={{ partitionLayout: PartitionLayout.sunburst, linkLabel: { maximumSection: 10000 } }}
     />
   </Chart>
 );