+++ storybook/stories/treemap/8_groove_text.story.tsx
@@ -9,8 +9,16 @@
 import { number } from '@storybook/addon-knobs';
 import React from 'react';
 
-import { Chart, Datum, MODEL_KEY, Partition, PartitionLayout, Settings } from '@elastic/charts';
-import { config } from '@elastic/charts/src/chart_types/partition_chart/layout/config';
+import {
+  Chart,
+  Datum,
+  MODEL_KEY,
+  Partition,
+  PartitionLayout,
+  Settings,
+  PartialTheme,
+  defaultPartitionValueFormatter,
+} from '@elastic/charts';
 import { ShapeTreeNode } from '@elastic/charts/src/chart_types/partition_chart/layout/types/viewmodel_types';
 import { arrayToLookup, hueInterpolator } from '@elastic/charts/src/common/color_calcs';
 import { mocks } from '@elastic/charts/src/mocks/hierarchical';
@@ -24,14 +32,25 @@ const countryLookup = arrayToLookup((d: Datum) => d.country, countryDimension);
 
 const interpolatorTurbo = hueInterpolator(palettes.turbo.map(([r, g, b]) => [r, g, b, 0.7]));
 
+const theme: PartialTheme = {
+  chartMargins: { top: 0, left: 0, bottom: 0, right: 0 },
+  partition: {
+    minFontSize: 4,
+    maxFontSize: 114,
+    idealFontSizeJump: 1.01,
+    outerSizeRatio: 1,
+  },
+};
+
 export const Example = () => (
   <Chart>
-    <Settings baseTheme={useBaseTheme()} />
+    <Settings theme={theme} baseTheme={useBaseTheme()} />
     <Partition
       id="spec_1"
       data={mocks.sunburst}
+      layout={PartitionLayout.treemap}
       valueAccessor={(d: Datum) => d.exportVal as number}
-      valueFormatter={(d: number) => `$${config.fillLabel.valueFormatter(Math.round(d / 1000000000))}\u00A0Bn`}
+      valueFormatter={(d: number) => `$${defaultPartitionValueFormatter(Math.round(d / 1000000000))}\u00A0Bn`}
       layers={[
         {
           groupByRollup: (d: Datum) => countryLookup[d.dest].continentCountry.slice(0, 2),
@@ -57,7 +76,7 @@ export const Example = () => (
           groupByRollup: (d: Datum) => d.dest,
           nodeLabel: (d: any) => countryLookup[d].name,
           fillLabel: {
-            valueFormatter: (d: number) => `${config.fillLabel.valueFormatter(Math.round(d / 1000000000))}\u00A0Bn`,
+            valueFormatter: (d: number) => `${defaultPartitionValueFormatter(Math.round(d / 1000000000))}\u00A0Bn`,
             textColor: 'black',
             fontWeight: 200,
             fontStyle: 'normal',
@@ -83,14 +102,6 @@ export const Example = () => (
           },
         },
       ]}
-      config={{
-        partitionLayout: PartitionLayout.treemap,
-        margin: { top: 0, bottom: 0, left: 0, right: 0 },
-        minFontSize: 4,
-        maxFontSize: 114,
-        idealFontSizeJump: 1.01,
-        outerSizeRatio: 1,
-      }}
     />
   </Chart>
 );